+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Fixed `Recipient.existsTypedProperty` and `AttachmentBase.existsTypedProperty` having the wrong return type.
 * Removed "TODO" markers on `OleStreamStruct` and finalized it to only handle the OLEStream for embedded objects.
 * Fixed type annotations for `extract_msg.utils.fromTimeStamp`.
+* Added new function `extract_msg.properties.prop.createNewProp`. This function allows a new property to be created with default data based on the name. The name MUST be an 8 character hex string, the first 4 characters being the value for the property ID and the second 4 being the value for the type.
 
 **v0.46.2**
 * Adjusted typing information on regular expressions. They were using a subscript that was added in Python 3.9 (apparently that is something the type checker doesn't check for), which made the module incompatible with Python 3.8. If you are using Python 3.9 or higher a version check will switch to the more specific typing.